Added extensible code for store

+++ rate_limiter_system/server/config.go
@@ -1,5 +1,5 @@
 package server
 
 type IPRateLimitMappingConfig struct {
-	IPRateLimits map[string][]int `json:"ip_rate_limits"`
+	IPRateLimits map[string][]int64 `json:"ip_rate_limits"`
 }
